SimplePIR CUDA acceleration proofs
We add a couple kernels to test how to maximize usage of the GPU.
The most basic kernel performs worse than a CPU as we have one CUDA thread per row in the mat-vec multiply.
We add coalesced loads, usage of shared memory, and warp/block/SM-aware ideas to reach a final kernel. This kernel achieves throughput of 177 GB/s, which is very close to the performance of the kernel that only reads A from memory on a warp-basis (185 GB/s).